Update Ejercicio 21.py

+++ Ejercicios con condicionales/Ejercicio 21.py
@@ -1,8 +1,8 @@
 #Programa que calcula una ecuación de segundo grado. Controla que el valor de la raíz cuadrada no de un valor negativo
 import math
-a= int(input("Introduce el valor de a: "))
-b= int(input("Introduce el valor de b: "))
-c= int(input("Introduce el valor de c: "))
+a=float(input("Introduce el valor de a: "))
+b=float(input("Introduce el valor de b: "))
+c=float(input("Introduce el valor de c: "))
 número_del_cual_sacamos_raíz= b**2 - 4*a*c
 if número_del_cual_sacamos_raíz<0:
     print("La ecuación no tiene solución real")
